feat(effects/v3): descending counterparts for Range and Ascending
An `IdList` could only compress upward. A strictly descending list — every id
its own segment — was the one shape the format had no answer for, and the test
that recorded it said so: `a_non_ascending_list_never_collapses` asserted 1,000
segments for 1,000 ids.

The shape that produces it is `CREATE_EDGE`'s `src` and `dst`. Those follow
*edge* id order, so their contents are unordered with respect to themselves: a
scan that walks nodes downward writes a strictly descending endpoint column.

Two new segments, mirroring the two that existed:

    RangeDescending { base, len }   ids base, base-1, ... base-len+1
    Descending { bitmap, len, min } the same roaring set, read from the top

**Direction is a header bit, not two more kinds.** Bit 6 of the segment header
is `SEG_DESCENDING`; the reserved mask narrows to bit 7 alone. A descending
segment is the same payload read the other way — `Range` gains a
first-and-highest base instead of a first-and-lowest one, and a bitmap is a
*set*, which has no direction at all. Two more kinds would have meant a second
blob format and a second collapse cost model for encodings that are
byte-identical. `direction_is_the_header_bit_and_the_blob_is_the_same_bytes`
pins that: the ascending and descending forms of the same 10,000 ids differ in
exactly one bit and agree on every following byte.

`Repeat` has no direction — one id, however many times — so the bit is refused
there rather than ignored, and a descending range that would step past zero is
refused the way an ascending one that would wrap past `u64::MAX` already was.

The collapse arithmetic is untouched. `Run` learns a `desc: Option<bool>`,
`None` until a second segment settles it, because a lone id belongs to neither
direction and the id after it is what decides. `absorb` now takes the segment
rather than a rebuilt ascending one, since a descending base is the run's
highest id and can need a wider value field. What ends a run is a *reversal*
rather than any step backwards.

WIRE CHANGE. A v3 decoder without this rejects a descending segment as a
reserved bit, which is the loud failure that mask exists for, but it means the
C implementation and docs/effects-v3.md both need the bit before anything emits
one. Nothing emits one until an endpoint column actually descends, so the
format is inert until C catches up — but it is a change to a format that was
frozen, and should be called out in #2698 and #2706.

+++ graph/src/effects/v3/value.rs
@@ -284,7 +284,7 @@ fn read_one(
         si_type::T_DATE => Value::Date(r.i64()?),
         si_type::T_TIME => Value::Time(r.i64()?),
         si_type::T_DURATION => Value::Duration(r.i64()?),
-        other => return Err(DecodeError::BadValueType(other as u32)),
+        other => return Err(DecodeError::BadValueType(other)),
     };
     Ok(Some(v))
 }
